refactor(authup): the commands become citty definitions over concern folders
`apps/authup/src` already mirrors server-core's `cli/` shape, but the
substance sat in the command files rather than beside them. `commands/dev.ts`
held eight top-level symbols, `commands/start.ts` defined two commands, and
the console-path guard existed twice: once as `assertConsolePath` in the dev
command, once written inline in `start`.

What server-core does is not more folders, it is that a command file is small
BECAUSE its concern has a home. So:

- `roles/` becomes `console/`, which is what it always held. `readConsoleConfigs`
  moves there next to `buildConsoleApplications` (out of `start`) and the one
  `assertConsolePath` both callers now share. The two error messages were
  already identical once the caller's name is spelled the same way, so the
  wording is unchanged.
- The dev machinery moves INTO `src/dev/`, never a `commands/dev/` that would
  sit confusingly beside it: `mount.ts` (compose plus the two mount builders),
  `constants.ts`, `env.ts`, and `open-in-editor.ts`, which earns its own file
  because it is a security guard rather than part of the vite server factory.
- `start` and `core` are one file each, sharing the factory that carries the
  reasoning for both roles.

Pure relocation: no behaviour changes, and the suite passes with one import
path edited.
