+++ apps/authup/src/commands/index.ts
@@ -7,6 +7,7 @@
 
 export * from './config.ts';
 export * from './console.ts';
+export * from './core.ts';
 export * from './dev.ts';
 export * from './healthcheck.ts';
 export * from './start.ts';
