+++ apps/authup/src/commands/start.ts
@@ -5,8 +5,6 @@
  * view the LICENSE file that was distributed with this source code.
  */
 
-import { AuthupError } from '@authup/errors';
-import { getURLBasePath } from '@authup/kit';
 import { defineCommand } from 'citty';
 import type { ConfigReadFsOptions } from '@authup/server-config';
 import {
@@ -16,90 +14,9 @@ import {
     readConfig,
     registerShutdownHandlers,
 } from '@authup/server-core';
-import type { ConsoleConfigs } from '../roles/config.ts';
-import { readConsoleConfigs } from '../roles/config.ts';
-import { createApplication as createAuthConsoleApplication } from '@authup/server-auth-console';
-import { createApplication as createAdminConsoleApplication } from '@authup/server-admin-console';
-import { createApplication as createAccountConsoleApplication } from '@authup/server-account-console';
 import { InjectionKey } from '@authup/server-console-kit';
-import type { Application } from 'orkos';
-
-/**
- * The single-container composition (plan 101 D2): every enabled console on
- * server-core's own listener.
- *
- * Each console is set up as the APPLICATION it is, with `listen: false`, and
- * what gets mounted is the app its own graph built. The alternative -- asking
- * each console for a bare handler -- would run neither its config module nor
- * its theme module, so `authup start` would be a third way to start a console
- * that only resembles the two supported ones. A console owns its listener
- * everywhere except here, and here it is composed rather than reduced.
- *
- * The mount PATH is the path component of the console's url, never the url
- * itself. A console url is where a BROWSER reaches the console, so it carries
- * the origin the proxy publishes; the listener only ever sees the path.
- */
-async function buildConsoleApplications(consoles: ConsoleConfigs) : Promise<{
-    path: string,
-    application: Application,
-}[]> {
-    const candidates : {
-        name: string,
-        url: string,
-        create: () => Application
-    }[] = [
-        // The auth console is not gated. The hosted login, consent and
-        // workflow pages are the issuance surface, so no flag turns them off
-        // (plan 099).
-        {
-            name: 'auth console',
-            url: consoles.auth.url,
-            create: () => createAuthConsoleApplication({ config: consoles.auth, listen: false }),
-        },
-    ];
-
-    if (consoles.admin.enabled) {
-        candidates.push({
-            name: 'admin console',
-            url: consoles.admin.url,
-            create: () => createAdminConsoleApplication({ config: consoles.admin, listen: false }),
-        });
-    }
-
-    if (consoles.account.enabled) {
-        candidates.push({
-            name: 'account console',
-            url: consoles.account.url,
-            create: () => createAccountConsoleApplication({ config: consoles.account, listen: false }),
-        });
-    }
-
-    const applications : { path: string, application: Application }[] = [];
-
-    for (const candidate of candidates) {
-        const path = getURLBasePath(candidate.url);
-        if (!path) {
-            // A console with no path of its own would have to own the API's
-            // root, where it would shadow the protocol routes and the page
-            // GETs would redirect to themselves. Refuse it by name rather
-            // than booting into a redirect loop. The origin needs no check
-            // here: every console url is already refused unless it is
-            // publicUrl's own origin, by the key that resolves it.
-            throw new AuthupError(
-                `The ${candidate.name} url is ${candidate.url}, which is this deployment's own origin root. ` +
-                'A console needs a path of its own; the defaults are under /console.',
-            );
-        }
-
-        const application = candidate.create();
-
-        await application.setup();
-
-        applications.push({ path, application });
-    }
-
-    return applications;
-}
+import type { ConsoleApplication } from '../console/index.ts';
+import { buildConsoleApplications, readConsoleConfigs } from '../console/index.ts';
 
 /**
  * The two roles that run server-core in this process, differing only in what
@@ -120,7 +37,7 @@ async function buildConsoleApplications(consoles: ConsoleConfigs) : Promise<{
  * argument at the registration site would say nothing about which role it
  * selects.
  */
-function defineApplicationCommand(
+export function defineApplicationCommand(
     configFs: ConfigReadFsOptions,
     role: { name: string, consoles: boolean },
 ) {
@@ -132,7 +49,7 @@ function defineApplicationCommand(
                 fs: { ...configFs },
             });
 
-            const consoles : { path: string, application: Application }[] = [];
+            const consoles : ConsoleApplication[] = [];
 
             const app = createApplication({
                 config: new ConfigModule(config),
@@ -174,8 +91,4 @@ function defineApplicationCommand(
 
 export function defineCLIStartCommand(configFs: ConfigReadFsOptions = {}) {
     return defineApplicationCommand(configFs, { name: 'start', consoles: true });
-}
-
-export function defineCLICoreCommand(configFs: ConfigReadFsOptions = {}) {
-    return defineApplicationCommand(configFs, { name: 'core', consoles: false });
 }
