+++ apps/authup/src/console/index.ts
@@ -1,0 +1,11 @@
+/*
+ * Copyright (c) 2026.
+ * Author Peter Placzek (tada5hi)
+ * For the full copyright and license information,
+ * view the LICENSE file that was distributed with this source code.
+ */
+
+export * from './applications.ts';
+export * from './config.ts';
+export * from './path.ts';
+export * from './types.ts';
