+++ apps/authup/src/dev/index.ts
@@ -7,6 +7,10 @@
 
 export * from './console-auth.ts';
 export * from './console-static.ts';
+export * from './constants.ts';
+export * from './env.ts';
+export * from './mount.ts';
+export * from './open-in-editor.ts';
 export * from './rate-limit.ts';
 export * from './resolve.ts';
 export * from './server.ts';
